Not allow already settled class to be deleted

+++ src/class/application/commands/handlers/delete-class-by-id.handler.ts
@@ -3,6 +3,7 @@ import { DeleteClassByIdCommand } from '../impl';
 import { ClassRepository } from 'src/class/infrastructure/repositories'
 import { BroadcastService, ClassDeletedEvent, ClassDeletedEventPayload } from '@tutorify/shared';
 import { Builder } from 'builder-pattern';
+import { BadRequestException, NotFoundException } from '@nestjs/common';
 
 @CommandHandler(DeleteClassByIdCommand)
 export class DeleteClassByIdHandler implements ICommandHandler<DeleteClassByIdCommand> {
@@ -13,6 +14,13 @@ export class DeleteClassByIdHandler implements ICommandHandler<DeleteClassByIdCo
 
     async execute(query: DeleteClassByIdCommand) {
         const { id } = query;
+        const classToDelete = await this.classRepository.findOneBy({id});
+        if (!classToDelete) {
+            throw new NotFoundException(`Class with id ${id} not exist`);
+        }
+        if (classToDelete.tutorId) {
+            throw new BadRequestException(`Class is settled! If you mean to cancel this class, contact the manager of the system`);
+        }
         await this.classRepository.delete(id);
         this.dispatchEvent(id);
         return true;
